+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## x.x.x - x-x-x
+
+**Patches**
+- Fixed codegen error when using React Native 0.74+ with old architecture by converting EventEmitter properties to callback functions in TurboModule interface. [#1977](https://github.com/stripe/stripe-react-native/issues/1977)
+
+
 ## 0.49.0 - 2025-07-02
 
 **Features**